Documentation fixes based on review. Added section for Part Choice operation Moved query operation models to different folder
