+++ CustomPLMService/PlmServiceImpl.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using Altium.PLM.Custom;
 using AutoMapper;
 using CustomPLMService.Contract;
+using CustomPLMService.Contract.Models;
 using CustomPLMService.Contract.Models.Authentication;
 using Grpc.Core;
 using Microsoft.Extensions.Logging;
@@ -36,7 +37,7 @@ using Id = CustomPLMService.Contract.Models.Items.Id;
 using ItemCreateSpec = CustomPLMService.Contract.Models.Items.ItemCreateSpec;
 using ItemUpdateSpec = CustomPLMService.Contract.Models.Items.ItemUpdateSpec;
 using OperationSupportedResponseTO = Altium.PLM.Custom.OperationSupportedResponse;
-using Query = CustomPLMService.Contract.Models.Search.Query;
+using Query = CustomPLMService.Contract.Models.Query.Query;
 using RelationshipTable = CustomPLMService.Contract.Models.Relationship.RelationshipTable;
 using RelationshipTableTO = Altium.PLM.Custom.RelationshipTable;
 using RelationshipType = CustomPLMService.Contract.Models.Metadata.RelationshipType;
