+++ CustomPLMService/PlmServiceMappingProfile.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using Altium.PLM.Custom;
 using AutoMapper;
 using AutoMapper.Internal;
+using CustomPLMService.Contract.Models;
 using CustomPLMService.Contract.Models.Authentication;
 using Google.Protobuf.Collections;
 using AttributeSpec = CustomPLMService.Contract.Models.Metadata.AttributeSpec;
@@ -41,7 +42,7 @@ using NumberingFormat = CustomPLMService.Contract.Models.Items.NumberingFormat;
 using RelationshipTypeTO = Altium.PLM.Custom.RelationshipType;
 using RelationshipSpecTO = Altium.PLM.Custom.RelationshipSpec;
 using NumberingFormatTO = Altium.PLM.Custom.NumberingFormat;
-using Query = CustomPLMService.Contract.Models.Search.Query;
+using Query = CustomPLMService.Contract.Models.Query.Query;
 using RelationshipSpec = CustomPLMService.Contract.Models.Metadata.RelationshipSpec;
 using RelationshipTable = CustomPLMService.Contract.Models.Relationship.RelationshipTable;
 using RelationshipType = CustomPLMService.Contract.Models.Metadata.RelationshipType;
