+++ Documentation/images/ConnectorAsAdapter.png
