+++ FilesystemPLMDriver/FileSystemPlmService.cs
@@ -3,12 +3,13 @@ using CustomPLMService.Contract.Models.Authentication;
 using CustomPLMService.Contract.Models.Items;
 using CustomPLMService.Contract.Models.Metadata;
 using CustomPLMService.Contract.Models.Relationship;
-using CustomPLMService.Contract.Models.Search;
 using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
+using CustomPLMService.Contract.Models;
+using CustomPLMService.Contract.Models.Query;
 using FilesystemPLMDriver.Models;
 using Microsoft.Extensions.Logging;
 using Type = CustomPLMService.Contract.Models.Metadata.Type;
