+++ FilesystemPLMDriver/appsettings.json
@@ -46,9 +46,9 @@
       "MCO"
     ],
     "AttributeNames": [
-      "number",
-      "description",
-      "name",
+      "Number",
+      "Description",
+      "Name",
       "Product Line",
       "Description",
       "Mount",
@@ -61,6 +61,12 @@
       "Value",
       "Power",
       "Current",
+      "MFR#1",
+      "MFR#2",
+      "MFR#3",
+      "MPN#1",
+      "MPN#2",
+      "MPN#3",
       "refDes"
     ]
   }
